+++ webpack.config.js
@@ -1,38 +1,73 @@
+const { exec } = require('child_process')
 const path = require('path')
 const webpack = require('webpack')
+const { ESBuildMinifyPlugin } = require('esbuild-loader')
 
-const babelOptions = {
-  // For debugging, just remove "@babel/preset-env":
-  presets: ['@babel/preset-env'],
-  plugins: [['@babel/plugin-transform-for-of', { assumeArray: true }]],
-  cacheDirectory: true
+const debug = process.env.WEBPACK_SERVE
+
+// Try exposing our socket to adb (errors are fine):
+if (process.env.WEBPACK_SERVE) {
+  console.log('adb reverse tcp:8083 tcp:8083')
+  exec('adb reverse tcp:8083 tcp:8083', () => {})
 }
 
+const bundlePath = path.resolve(
+  __dirname,
+  'android/src/main/assets/edge-exchange-plugins'
+)
+
 module.exports = {
-  devtool: 'source-map',
-  entry: './src/react-native.js',
-  mode: 'development',
+  devtool: debug ? 'source-map' : undefined,
+  devServer: {
+    allowedHosts: 'all',
+    hot: false,
+    port: 8083,
+    static: bundlePath
+  },
+  entry: './src/index.ts',
+  mode: debug ? 'development' : 'production',
   module: {
     rules: [
       {
-        test: /\.js$/,
-        use: { loader: 'babel-loader', options: babelOptions }
+        exclude: /\/node_modules\//,
+        test: /\.ts$/,
+        use: {
+          loader: 'esbuild-loader',
+          options: { loader: 'ts' }
+        }
       }
     ]
   },
+  optimization: {
+    minimizer: [
+      new ESBuildMinifyPlugin({
+        target: 'chrome67'
+      })
+    ]
+  },
   output: {
+    chunkFilename: '[name].chunk.js',
     filename: 'edge-exchange-plugins.js',
-    path: path.join(path.resolve(__dirname), 'lib/react-native')
+    path: bundlePath
   },
   plugins: [
-    new webpack.ProvidePlugin({ Buffer: ['buffer', 'Buffer'] }),
-    new webpack.ProvidePlugin({ process: ['process'] })
+    new webpack.IgnorePlugin({ resourceRegExp: /^(https-proxy-agent)$/ }),
+    new webpack.ProvidePlugin({
+      Buffer: ['buffer', 'Buffer']
+    }),
+    new webpack.ProvidePlugin({
+      process: path.resolve('node_modules/process/browser.js')
+    })
   ],
   resolve: {
+    extensions: ['.ts', '.js'],
     fallback: {
-      assert: require.resolve('assert/'),
-      buffer: require.resolve('buffer/'),
-      stream: require.resolve('stream-browserify')
+      crypto: require.resolve('crypto-browserify'),
+      fs: false,
+      http: require.resolve('stream-http'),
+      https: require.resolve('https-browserify'),
+      stream: require.resolve('stream-browserify'),
+      url: require.resolve('url')
     }
   },
   target: ['web', 'es5']